# platform/scenario-b/apis/openapi/amm.yaml
openapi: 3.0.3
info:
  title: CBWeb3 Platform — AMM API (Scenario B)
  version: "2.0.0"
  description: |
    API Gateway endpoints for the Automated Market Maker (Scenario B).
    Covers cooperative liquidity provisioning, commit-reveal, swaps, quotes,
    and multi-pair PairRegistry management (005-cooperative-liquidity).
servers:
  - url: /api/v2/amm
    description: Scenario B AMM namespace
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
...
paths:
...
  /liquidity/commit:
    post:
      summary: Register a single-sided liquidity commit (commit phase)
      operationId: commitLiquidity
      tags: [Liquidity]
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pool_pair, provider_id, side, amount]
              properties:
                pool_pair:
                  type: string
                  example: "BRL-USD"
                provider_id:
                  type: string
                side:
                  type: string
                  enum: [A, B]
                amount:
                  type: string
      responses:
        "201":
          description: Commit registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  commit_id:
                    type: string
                  status:
                    type: string
        "422":
          description: Commit failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /liquidity/commits:
    get:
      summary: List liquidity commits
      operationId: listCommits
      tags: [Liquidity]
      security:
        - cookieAuth: []
      parameters:
        - name: pool_pair
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of commits
          content:
            application/json:
              schema:
                type: object
                properties:
                  commits:
                    type: array
                    items:
                      type: object
  /liquidity/commits/{commit_id}:
    delete:
      summary: Cancel a pending commit
      operationId: cancelCommit
      tags: [Liquidity]
      security:
        - cookieAuth: []
      parameters:
        - name: commit_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Commit cancelled
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /liquidity/add:
    post:
      summary: Add liquidity (dual-sided — legacy)
      operationId: addLiquidity
      tags: [Liquidity]
      description: |
        Legacy dual-sided liquidity provisioning endpoint.
        Used by tryouts and governance frontend for cooperative scenarios.
        For sovereign CB flow, use the commit-reveal + bridge-based approach instead.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pool_pair, provider_bank_id, token_a_amount, token_b_amount]
              properties:
                pool_pair:
                  type: string
                provider_bank_id:
                  type: string
                token_a_amount:
                  type: string
                token_b_amount:
                  type: string
      responses:
        "201":
          description: Liquidity added
        "422":
          description: Failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /liquidity/remove:
    post:
      summary: Remove liquidity (proportional withdrawal)
      operationId: removeLiquidity
      tags: [Liquidity]
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pool_pair, provider_bank_id, lp_id]
              properties:
                pool_pair:
                  type: string
                provider_bank_id:
                  type: string
                lp_id:
                  type: string
      responses:
        "200":
          description: Liquidity removed
        "422":
          description: Failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ── Token Mint + Approve ──────────────────────────────────────────────
...